requirements.txt
requests 
urllib3 
PyQt5==5.15.6  
PyQtWebEngine==5.15.6
pyinstaller>=5.6.2
pillow>=9.3.0
